Allow to specify config file path
Now it is possible to specify which file to grab the config from.
It might be a JSON or JS file and also the "chimi" field in the package.json.

Made all config values passed as parameters to the functions using them,
which removes the implicit dependency on the config file inside the lib files,
this allows to use chimi programatically.